requirements.txt
asgiref==3.6.0
Django==3.2
djangorestframework==3.13.1
pytz==2022.7.1
sqlparse==0.4.3
django-filter==21.1